requirements.txt
streamlit == 1.34.0
pandas == 2.2.2
# plotting libraries
matplotlib == 3.9.0
seaborn == 0.13.2
# image library
pillow == 10.3.0
# library for analysis
numpy == 1.26.4
statistics == 1.0.3.5
streamlit-elements == 0.10.*
plotly-express == 0.4.1
scikit-learn == 1.5.0
Xlswriter == 3.2.0
fpdf == 1.7.2